my_app/requirements.txt
pandas
plotly
python-dotenv
Requests
streamlit
streamlit_extras
